requirements/development.txt
-r base.txt
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.5.1
django-debug-toolbar==4.3.0
